online_inference/requirements.txt
click==7.1.2
fastapi==0.65.1
h11==0.12.0
joblib==1.0.1
numpy==1.20.3
pandas==1.2.4
pydantic==1.8.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
starlette==0.14.2
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
uvicorn==0.13.4
yml==0.0.1
requests~=2.25.1
pytest~=6.2.4